Move repository implementation into its own file
